sake自体を渡していくようにする
